requirements.txt
flask
gunicorn
dnspython
pendulum
OTXv2
pyyaml